Split Python tests out of botan2.py
No reason to ship these to the end user
